+++ esibd/controls/__init__.py
@@ -1,0 +1,3 @@
+"""Label this directory as a Python package."""
+
+from .webcam.webcam import Webcam  # noqa: F401